refactor: oralce context
