+++ src/components/providers/ClientProviders.tsx
@@ -2,7 +2,6 @@
 
 import type { ReactNode } from 'react';
 import { GlobalModalProvider } from '@/contexts/GlobalModalContext';
-import { OracleDataProvider } from '@/contexts/OracleDataContext';
 import { OnboardingProvider } from '@/features/positions/components/onboarding/onboarding-context';
 
 type ClientProvidersProps = {
@@ -12,9 +11,7 @@ type ClientProvidersProps = {
 export function ClientProviders({ children }: ClientProvidersProps) {
   return (
     <GlobalModalProvider>
-      <OracleDataProvider>
-        <OnboardingProvider>{children}</OnboardingProvider>
-      </OracleDataProvider>
+      <OnboardingProvider>{children}</OnboardingProvider>
     </GlobalModalProvider>
   );
 }
